feat(): re-organize directory structure
- Add `SomeValidator` that instantiate proxy validator
